// dashboard/src/utils/hasura-api/generated/schemas/computedFieldItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { ComputedFieldDefinition } from './computedFieldDefinition';

export interface ComputedFieldItem {
  /** Name of the computed field */
  name: string;
  /** The computed field definition */
  definition: ComputedFieldDefinition;
  /** Customize the description shown in GraphQL introspection. If null or omitted then if a comment exists on the database function, it is used as the description, and if not, an autogenerated description is used instead. */
  comment?: string;
}